+++ src/extension.ts
@@ -1,11 +1,9 @@
 import * as vscode from 'vscode';
-
 import { AstTreeDataProvider } from './AstTreeDataProvider'
 
 export function activate(context: vscode.ExtensionContext) {
-
 	const tsAstOutlineProvider = new AstTreeDataProvider(context);
-	
+
 	vscode.commands.registerCommand('tsAstOutline.refresh', () => tsAstOutlineProvider.refresh());
 	vscode.commands.registerCommand('tsAstOutline.changeMode', () => tsAstOutlineProvider.changeMode());
 	vscode.commands.registerCommand('tsAstOutline.refreshNode', node => tsAstOutlineProvider.refresh(node));
@@ -14,22 +12,21 @@ export function activate(context: vscode.ExtensionContext) {
 	vscode.commands.registerCommand('tsAstOutline.removeNode', node => tsAstOutlineProvider.removeNode(node));
 	vscode.commands.registerCommand('tsAstOutline.addChild', node => tsAstOutlineProvider.addChild(node));
 	vscode.commands.registerCommand('extension.openJsonSelection', range => tsAstOutlineProvider.select(range));
-
 }
-
-
-export interface Settings {
-	autoRefresh: boolean
-}
-
-export interface ProjectManagerOptions { // change name to something more like Status or State... a
-  mode: 'getChildren' | 'forEachChildren',
-  // showSourceFiles: boolean
-}
-
 
 export function readSettings(): Settings {
 	return {
 		autoRefresh: vscode.workspace.getConfiguration('tsAstOutline').get('autorefresh')||false
 	}
 }
+
+export interface Settings {
+	autoRefresh: boolean
+}
+
+export interface State { 
+  mode: 'getChildren' | 'forEachChildren'
+	/** user can toggle auto-refresh */
+	autoRefresh: boolean
+}
+